fix the mangling only in one place (where Twisted requires it)
