+++ src/klein/test/test_headers.py
@@ -23,11 +23,6 @@ from .._headers import (
     normalizeHeaderName, normalizeRawHeaders, normalizeRawHeadersFrozen,
 )
 
-try:
-    from twisted.web.http_headers import _sanitizeLinearWhitespace
-except ImportError:
-    _sanitizeLinearWhitespace = None
-
 # Silence linter
 AnyStr, Dict, Iterable, List, Optional, RawHeaders, Text, Tuple
 
@@ -54,18 +49,6 @@ def decodeName(name):
 def decodeValue(name):
     # type: (bytes) -> Text
     return name.decode(HEADER_VALUE_ENCODING)
-
-
-def _twistedHeaderNormalize(value):
-    # type: (Text) -> Text
-    """
-    Normalize the given header value according to the rules of the installed
-    Twisted version.
-    """
-    if _sanitizeLinearWhitespace is None:
-        return value
-    else:
-        return _sanitizeLinearWhitespace(value.encode("utf-8")).decode("utf-8")
 
 
 class EncodingTests(TestCase):
@@ -256,6 +239,15 @@ class GetValuesTestsMixIn(object):
             )
 
 
+    def headerNormalize(self, value):
+        # type: (Text) -> Text
+        """
+        Test hook for the normalization of header text values, which is a
+        behavior Twisted has changed after version 18.9.0.
+        """
+        return value
+
+
     @given(iterables(tuples(ascii_text(min_size=1), latin1_text())))
     def test_getTextName(self, textPairs):
         # type: (Iterable[Tuple[Text, Text]]) -> None
@@ -280,7 +272,7 @@ class GetValuesTestsMixIn(object):
         for name, _values in textValues.items():
             cast(TestCase, self).assertEqual(
                 list(self.getValues(rawHeaders, name)),
-                [_twistedHeaderNormalize(value) for value in _values],
+                [self.headerNormalize(value) for value in _values],
                 "header name: {!r}".format(name)
             )
 
@@ -311,7 +303,7 @@ class GetValuesTestsMixIn(object):
         for textName, values in binaryValues.items():
             cast(TestCase, self).assertEqual(
                 tuple(self.getValues(rawHeaders, textName)),
-                tuple(_twistedHeaderNormalize(headerValueAsText(value))
+                tuple(self.headerNormalize(headerValueAsText(value))
                       for value in values),
                 "header name: {!r}".format(textName)
             )
